crew_ai/memory_and_external_tools/pyproject.toml
[project]
name = "memory_and_external_tools"
version = "0.1.0"
description = "memory_and_external_tools using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai>=0.102.0",
    "google-generativeai",
    "mem0ai>=0.1.66",
    "pydantic>=2.10.6",
]

[project.scripts]
memory_and_external_tools = "memory_and_external_tools.main:run"
memory_project = "memory_and_external_tools.main:kickoff"
memory_project_plot = "memory_and_external_tools.main:plot"

run_crew = "memory_and_external_tools.main:run"
train = "memory_and_external_tools.main:train"
replay = "memory_and_external_tools.main:replay"
test = "memory_and_external_tools.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
